requirements.txt
python-telegram-bot==20.3
python-dotenv
youtube-search-python
pytube
requests